refactor(resources/messages): refactor messages
